Change textfield emoji input logic.
