+++ xDrip/Resources/en.lproj/Localizable.strings
@@ -43,6 +43,7 @@
 "edit_food_entry_bolus_section_footer" = "Choose bolus units amount and select the precise date of the injection";
 "edit_food_entry_carbs_amount_title" = "Amount consumed";
 "edit_food_entry_type_of_food" = "Type of food";
+"edit_food_entry_type_of_food_textfield_placeholder" = "Enter food type...";
 "edit_food_entry_bolus_amount_title" = "Units";
 "edit_food_entry_date_and_time" = "Date/Time";
 "edit_food_entry_carbs_amount_unit_grams" = "g";
