+++ xDripTests/Scenes/Settings/BaseSettings/BaseSettingsViewControllerTests.swift
@@ -89,7 +89,13 @@ final class BaseSettingsViewControllerTests: XCTestCase {
             .disclosure(mainText: "", detailText: nil, selectionHandler: {}),
             .pickerExpandable(mainText: "", detailText: nil, picker: CustomDatePicker()),
             .rightSwitch(text: "", isSwitchOn: false, switchHandler: { _ in }),
-            .textInput(mainText: "", detailText: nil, placeholder: nil, textChangedHandler: { _ in }),
+            .textInput(
+                mainText: "",
+                detailText: nil,
+                textFieldText: nil,
+                placeholder: nil,
+                textChangedHandler: { _ in }
+            ),
             .volumeSlider(value: 0.0, changeHandler: { _ in })
         ]
         
@@ -112,7 +118,13 @@ final class BaseSettingsViewControllerTests: XCTestCase {
             .disclosure(mainText: "", detailText: nil, selectionHandler: {}),
             .pickerExpandable(mainText: "", detailText: nil, picker: CustomDatePicker()),
             .rightSwitch(text: "", isSwitchOn: false, switchHandler: { _ in }),
-            .textInput(mainText: "", detailText: nil, placeholder: nil, textChangedHandler: { _ in }),
+            .textInput(
+                mainText: "",
+                detailText: nil,
+                textFieldText: nil,
+                placeholder: nil,
+                textChangedHandler: { _ in }
+            ),
             .volumeSlider(value: 0.0, changeHandler: { _ in })
         ]
         
@@ -160,7 +172,13 @@ final class BaseSettingsViewControllerTests: XCTestCase {
             .disclosure(mainText: "", detailText: nil, selectionHandler: {}),
             .pickerExpandable(mainText: "", detailText: nil, picker: CustomDatePicker()),
             .rightSwitch(text: "", isSwitchOn: false, switchHandler: { _ in }),
-            .textInput(mainText: "", detailText: nil, placeholder: nil, textChangedHandler: { _ in }),
+            .textInput(
+                mainText: "",
+                detailText: nil,
+                textFieldText: nil,
+                placeholder: nil,
+                textChangedHandler: { _ in }
+            ),
             .volumeSlider(value: 0.0, changeHandler: { _ in })
         ]
